requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.0.1
boto3==1.38.1
botocore==1.38.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.18.0
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
intervaltree==3.1.0
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
mypy-boto3-dynamodb==1.38.0
mypy-boto3-ecs==1.38.1
mypy-boto3-s3==1.38.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
